AI_UseCase/requirements.txt
streamlit 
langchain-openai 
langchain-groq 
langchain-google-genai 
langchain-core
langchain-community
faiss-cpu
tavily-python
pypdf